refactor/이미지-관련-로직-최적화 노티피케이션 로직을 공통 모듈로 추출
